test(settings): split color conversion coverage
Why:
- The remaining ARGB color conversion tests in AppAndReader exercise BibleUI color-settings helpers, not app delegate or scene bootstrap behavior.
- App-host tests should be limited to coverage that genuinely needs the running app target.

What Changed:
- Moved the color byte-clamping and signed ARGB serialization tests into BibleUITests.
- Left AndBibleTests+AppAndReader with only the scene-configuration sentinel.
- Updated the migration plan to record the final AppAndReader boundary.

Validation:
- GitNexus impact checks for the moved test methods were LOW with no direct callers or affected flows.
- GitNexus detect_changes scope=staged reported no mapped production-flow changes.
- Ran focused BibleUI package tests for ColorARGBConversionTests.
- Ran focused app-host test for testApplicationDelegateSceneConfigurationUsesWindowSceneDelegate.
- Ran git diff --cached --check, docblock guard, and source-guard checks.
- Adversarial subagent review found no issues.

Impact:
- Moves the last package-owned AppAndReader tests into the app-host-free package lane.
- Keeps true app-bootstrap coverage in the app-host bundle without changing production behavior.
